+++ forge/src/main/java/com/binaris/wizardry/datagen/EBForgeDataGenerators.java
@@ -45,6 +45,7 @@ public class EBForgeDataGenerators {
         generator.addProvider(event.includeClient(), new EBItemModelProvider(packOutput, existingFileHelper));
         generator.addProvider(event.includeServer(), new EBSpellsProvider(packOutput));
         generator.addProvider(event.includeServer(), new EBSpellDocsProvider(packOutput));
+        generator.addProvider(event.includeServer(), new EBArtifactDocsProvider(packOutput));
 
         EBBlockTagProvider blockTagProvider = generator.addProvider(event.includeServer(), new EBBlockTagProvider(packOutput, lookupProvider, existingFileHelper));
         generator.addProvider(event.includeServer(), new EBItemTagProvider(packOutput, lookupProvider, blockTagProvider.contentsGetter(), existingFileHelper));